Fixed incorrect path for results page.
As given, the example puts results.html under templates.  However, in
WebController a "redirect:/results" is used, which will not work as
shipped.

We moved the results.html (which is raw html and not a thymeleaf
template) into the static directory and modified the redirect in
WebController.

Navya & Dirk, Pivotal Labs Santa Monica