Rename Person to PersonForm to make obvious it's not considered a domain type.
I've been pointed to this guide by a couple of people that use it as evidence that the Spring team advocates using JSR-303 annotations to validate domain types. I can see that you can get this impression from this guide but I hope that's not its intention. Using JSR-303 on domain objects is problematic for a couple of reasons:

It creates incentives to build a weak domain model in which types don't enforce their invariants themselves but rely on some external frameworks to enforce them. This is a bad idea as downstream code receiving instances of those types can never be sure the instance is in a valid state. In this particular example, assume a RegistrationService.register(Person person). The implementation of that code can never be sure about anything in the person instance as it can easily be called like this: `register(new Person()). The crucial point here is, that Person — if considered a domain type — should not be able to be instantiated in anything but a valid state (101 of class design).

However, JSR-303 has a decent role in implementing validation on classes that are particularly designed to take potentially bogus user input to collect the errors and return them to the user. Those types are usually referred to as form backing objects and it would be cool if the guide would make explicit that what's currently referred to as Person is in fact a PersonForm.

Sorry for being so verbose but I thought that even if a minor change, the implications and their justification are worth being captured for reference.